+++ two_factor_authentication/constants.py
@@ -1,0 +1,1 @@
+MSG_TPL = 'Two-factor verification {}'
