+++ two_factor_authentication/serializers.py
@@ -1,4 +1,6 @@
 from rest_framework import serializers
+
+from Web_Menu_DA import settings
 from registration.models import WebMenuUser
 from two_factor_authentication.models import GoogleAuth
 
@@ -22,8 +24,10 @@ class Enable2faSerializer(serializers.ModelSerializer):
 
 class GoogleAuthSerializer(serializers.ModelSerializer):
     owner = serializers.ReadOnlyField(source='owner.email')  # change visible info from owner_id to owner_email
+    # redirect_to = serializers.URLField(default=reverse('display_qr'))
+    # redirect_to = serializers.URLField(default=f'{reverse("display_qr")}')
+    redirect_to = serializers.URLField(default=f'{settings.HOST}/enable2fa/display_qr')
 
     class Meta:
-        ordering = ['-id']
         model = GoogleAuth
-        fields = ['id', 'owner', 'otp_base32', 'otp_auth_url']
+        fields = ['id', 'owner', 'otp_base32', 'otp_auth_url', 'redirect_to']
