--- two_factor_authentication/tests/system_errors.log
@@ -1,108 +1,0 @@
-2023-07-03 16:54:54,659 WARNING Bad Request: /enable2fa/
-2023-07-03 16:55:23,761 WARNING Bad Request: /enable2fa/
-2023-07-03 16:56:22,061 WARNING Bad Request: /enable2fa/
-2023-07-03 16:58:59,914 WARNING Bad Request: /enable2fa/
-2023-07-03 17:13:32,163 WARNING Bad Request: /enable2fa/
-2023-07-03 17:14:48,600 WARNING Bad Request: /enable2fa/
-2023-07-03 17:16:18,860 WARNING Bad Request: /enable2fa/
-2023-07-03 17:16:22,897 WARNING Bad Request: /enable2fa/
-2023-07-03 17:30:25,385 WARNING Bad Request: /enable2fa/
-2023-07-03 17:31:11,114 WARNING Bad Request: /enable2fa/
-2023-07-03 17:31:29,216 WARNING Bad Request: /enable2fa/
-2023-07-03 17:31:59,746 WARNING Bad Request: /enable2fa/
-2023-07-03 17:32:24,241 WARNING Bad Request: /enable2fa/
-2023-07-03 17:33:17,994 WARNING Bad Request: /enable2fa/
-2023-07-03 17:35:52,207 WARNING Bad Request: /enable2fa/
-2023-07-03 17:35:52,354 WARNING Bad Request: /enable2fa/
-2023-07-03 17:35:52,637 WARNING Bad Request: /enable2fa/
-2023-07-03 17:38:54,481 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:38:54,481 WARNING Bad Request: /company/
-2023-07-03 17:38:54,826 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:38:54,999 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:38:55,156 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:38:55,282 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:38:55,430 WARNING Unauthorized: /company/
-2023-07-03 17:38:55,550 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:38:55,691 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:39:43,233 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 17:39:43,233 WARNING Bad Request: /company/
-2023-07-03 17:41:57,388 WARNING Bad Request: /enable2fa/
-2023-07-03 17:41:57,544 WARNING Bad Request: /enable2fa/
-2023-07-03 17:41:57,827 WARNING Bad Request: /enable2fa/
-2023-07-03 17:42:16,839 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 18:17:28,879 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 18:17:28,879 WARNING Bad Request: /company/
-2023-07-03 18:27:28,856 WARNING we receive a request with a non-existent Types2FA
-2023-07-03 18:27:28,856 WARNING Bad Request: /company/
-2023-07-03 18:40:42,940 WARNING Bad Request: /company/
-2023-07-03 18:44:13,731 WARNING Bad Request: /enable2fa/
-2023-07-03 18:44:26,715 WARNING Bad Request: /enable2fa/
-2023-07-03 18:46:31,348 WARNING Bad Request: /enable2fa/
-2023-07-03 18:46:41,219 WARNING Bad Request: /enable2fa/
-2023-07-03 18:46:49,114 WARNING Bad Request: /enable2fa/
-2023-07-03 18:47:18,809 WARNING Bad Request: /company/
-2023-07-03 18:47:18,951 WARNING Bad Request: /company/
-2023-07-03 18:47:19,076 WARNING Bad Request: /company/
-2023-07-03 18:47:19,390 WARNING cache_code expired or problem with cache_set().
-2023-07-03 18:47:19,390 WARNING Bad Request: /company/
-2023-07-03 18:47:19,516 WARNING Unauthorized: /company/
-2023-07-03 18:47:19,656 WARNING Bad Request: /company/
-2023-07-03 18:47:36,465 WARNING Bad Request: /company/
-2023-07-03 18:47:42,779 WARNING Bad Request: /company/
-2023-07-03 18:47:56,157 WARNING cache_code expired or problem with cache_set().
-2023-07-03 18:47:56,157 WARNING Bad Request: /company/
-2023-07-03 18:48:02,462 WARNING Unauthorized: /company/
-2023-07-03 18:48:09,543 WARNING Bad Request: /company/
-2023-07-04 16:31:10,543 WARNING Method Not Allowed (POST): /enable2fa/display_qr
-2023-07-04 16:31:25,037 WARNING Method Not Allowed (POST): /enable2fa/display_qr
-2023-07-04 16:32:26,929 WARNING Method Not Allowed (POST): /enable2fa/display_qr
-2023-07-04 16:33:09,721 WARNING Method Not Allowed (POST): /enable2fa/display_qr
-2023-07-04 18:09:53,226 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:10:32,014 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:11:24,191 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:12:07,012 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:12:28,658 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:18:24,644 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:29:49,422 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:33:38,776 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:36:07,180 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:36:24,344 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:37:42,698 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:45:29,598 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:49:04,371 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:50:01,708 WARNING Not Found: /display_qr
-2023-07-04 18:51:39,807 WARNING Not Found: /display_qr
-2023-07-04 18:52:18,226 WARNING Not Found: /display_qr
-2023-07-04 18:53:00,846 WARNING Not Found: /display_qr
-2023-07-04 18:54:11,920 WARNING Not Found: /display_qr
-2023-07-04 18:59:14,747 WARNING Not Found: /display_qr
-2023-07-04 18:59:14,810 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 18:59:57,070 WARNING Not Found: /display_qr
-2023-07-04 18:59:57,126 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 19:00:19,847 WARNING Not Found: /display_qr
-2023-07-04 19:00:19,900 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 19:00:47,410 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 19:01:00,705 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 19:01:11,844 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 19:01:28,739 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 21:31:07,805 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 21:31:52,975 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 21:32:17,193 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 21:33:30,357 WARNING Not Found: /enable2fa/display_qr
-2023-07-04 21:33:59,674 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 16:17:02,428 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 16:17:26,040 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 16:20:11,389 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 17:18:17,044 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 17:38:05,077 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 17:44:16,480 WARNING Not Found: /enable2fa/display_qr
-2023-07-05 17:44:28,525 WARNING Bad Request: /company/
-2023-07-05 17:44:28,650 WARNING Bad Request: /company/
-2023-07-05 17:44:28,792 WARNING Bad Request: /company/
-2023-07-05 17:44:29,073 WARNING cache_code expired or problem with cache_set().
-2023-07-05 17:44:29,076 WARNING Bad Request: /company/
-2023-07-05 17:44:29,199 WARNING Unauthorized: /company/
-2023-07-05 17:44:29,325 WARNING Bad Request: /company/
-2023-07-05 17:44:38,105 WARNING Bad Request: /enable2fa/
-2023-07-05 17:44:38,247 WARNING Bad Request: /enable2fa/
-2023-07-05 17:44:38,482 WARNING Bad Request: /enable2fa/
